+++ tests/unit/test_filter_out_violations.py
@@ -22,8 +22,6 @@
 
 """Tests for ondivi."""
 
-import pytest
-
 from ondivi.filter_out_violations import filter_out_violations
 
 
@@ -33,6 +31,7 @@ def test_without_violation() -> None:
         {},
         ['All checks passed'],
         '{filename}:{line_num:d}:{col_num:d}: {message}',
+        only_violations=False,
     )
 
     assert violations == ['All checks passed']
@@ -45,6 +44,7 @@ def test_custom_format() -> None:
         {'file.py': [12]},
         ['line=12 file=file.py message=`print` found'],
         'line={line_num:d} file={filename} {other}',
+        only_violations=False,
     )
 
     assert violations == ['line=12 file=file.py message=`print` found']
@@ -57,6 +57,7 @@ def test_not_target_violation() -> None:
         {'file.py': [1, 2]},
         ['file.py:3:1: line too long'],
         '{filename}:{line_num:d}:{col_num:d}: {message}',
+        only_violations=False,
     )
 
     assert not violations
@@ -69,24 +70,24 @@ def test_file_without_diff() -> None:
         {'file.py': [1, 2]},
         ['foo.py:3:1: line too long'],
         '{filename}:{line_num:d}:{col_num:d}: {message}',
+        only_violations=False,
     )
 
     assert not violations
     assert not found
 
 
-# TODO #48 implement and remove `skip` flag
-@pytest.mark.skip()
 def test_only_violations() -> None:
     """Test only violations."""
     violations, found = filter_out_violations(
         {'file.py': [3]},
         [
-            'foo.py:3:1: line too long',
+            'file.py:3:1: line too long',
             'Info message',
         ],
         '{filename}:{line_num:d}:{col_num:d}: {message}',
+        only_violations=True,
     )
 
-    assert violations == ['foo.py:3:1: line too long']
+    assert violations == ['file.py:3:1: line too long']
     assert found